requirements.txt
cycler==0.10.0
future==0.18.2
futures==3.1.1
iso8601==0.1.12
kiwisolver==1.1.0
pyparsing==2.4.5
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
six==1.13.0
